(PUP-5016) Fix linux service test on Debian and Ubuntu
Previously, services_enable_linux.rb was failing in Debian 8
and Ubuntu 15.05 due to issues with Systemd and SysVInit. The
Debian 8 issues are corrected by the actual fix to the systemd
and init providers, but the test continued to fail in Ubuntu
due to an upstream systemd-sysv compatibility bug.

This commit updates the test to use a service which has
a systemd unit file, cron, rather than apache2 which did not. This
allows the test to run correctly on Ubuntu. Once this upstream bug
is fixed, we'll probably want to undo the Ubuntu changes for
consistency.

See https://bugs.launchpad.net/ubuntu/+source/systemd/+bug/1447807
for the upstream bug report.