(PUP-5317) Exit with 74 when trying to run a master with no puppet user
Commit 144ea123 changed Puppet's behavior when a master is run
with no Puppet user existing. This commit maintains the more
graceful behavior of failing in main rather than down the line,
but restores the behavior of exiting with `74`.

The test `puppet_manages_own_configuration_in_robust_manner` is
also updated to account for the fact that we no longer attempt
to change groups if the user is missing.