(PUP-5387) Wait for service to finish stopping in AIX before restarting
Previously when restarting a service, the AIX SRC provider would hit
the service with the stop signal via `stopsrc` followed by immediately
sending the start signal via `startsrc`. This was a problem for some
services for which stopping is not instant. In such a case, the service
would go into the 'stopping' state and ignore the start signal, causing
the restart to result in the service being stopped.

This commit adds a 30 second timeout during restarts to check that the
service is actually stopped before attempting to start it again.