(PUP-5387) Wait for service to finish stopping in AIX
Previously when stopping a service, the AIX SRC provider would hit
the service with the stop signal via `stopsrc` and return immediately
as the service asynchronously shut down. In some cases, this would
cause a race condition if the service took a moment to actually stop.
In such a case, Puppet became unaware of the actual state of the
service.

This commit adds a 60 second timeout when stopping services to check
that the service is actually stopped before returning.