(PUP-5233) Make `puppet inspect` use correct terminus class
Previously, the `puppet inspect` application always used the
:yaml indirection terminus. We now store cached catalogs in
json by default. Since we keep yaml and json catalogs in
separate locations, inspect was unable to locate a cached
catalog out of the box.

This commit updates the puppet inspect application to first use
the the catalog_cache_terminus setting to determine the
indirection terminus class. If it is not set, we will now
fall back to JSON by default.