(PUP-6159) Convert CFPropertyList binary strings to CFPropertyList::Blob's
Previously, Puppet would fail to set passwords in OSX via the
directoryservice provider due to CFPropertyList encoding problems
when converting entropy and salt strings into binary.

This was due to the fact that CFPropertyList was treating these
base64 encoded fields as CFStrings, which would then fail to
convert to binary due to invalid UTF-8 byte sequences.

In order to prevent this, the 'salt' and 'entropy' fields must
be converted into CFPropertyList::Blob's, which are interpreted
as CFPropertyList::CFData objects when parsed by CFPropertyList.
This is the correct data type for binary data values.

In addition, when setting shadow hash data, the binary_plist
string must also be converted into a Blob for the same reasons.